BACKEND/backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator
python-multipart
redis
